Updated citizen foreign key to properly map Integer to Integer

+++ api/app/models/citizen.py
@@ -25,7 +25,7 @@ class Citizen(Base):
     citizen_name = db.Column(db.String(150), nullable=True)
     citizen_comments = db.Column(db.String(1000), nullable=True)
     qt_xn_citizen_ind = db.Column(db.Integer, default=0, nullable=False)
-    cs_id = db.Column(db.BigInteger, db.ForeignKey('citizenstate.cs_id'), nullable=False)
+    cs_id = db.Column(db.Integer, db.ForeignKey('citizenstate.cs_id'), nullable=False)
     start_time = db.Column(db.DateTime, nullable=False)
 
     service_reqs = db.relationship('ServiceReq', lazy='joined')
